Tibia 8.42 Updates Refer to changelog for 1.19.5 - 1.19.7 for details

+++ sdk/TibiaCharacter.h
@@ -25,7 +25,7 @@ public:
 	int outfitId;
 	int monsterType;
 	int hpPercLeft;
-	int cap;
+	float cap;
 	int maxHp;
 	int maxMana;
 	int exp;
